+++ package.json
@@ -18,6 +18,8 @@
     "mongoose": "^4.11.12",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "request": "^2.81.0",
     "request-promise": "^4.2.1",
     "webpack": "^3.6.0"
