+++ public/Login.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export default class Login extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
+
+		}
+	}
+	render() {
+    return (
+  		<div>
+        <h2> Login </h2>
+  		</div>
+    )
+	}
+}
