+++ public/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx'
+import App from './App.jsx';
+import { browserHistory } from 'react-router';
 
 
-ReactDOM.render(<App/>, document.getElementById('app'));
+ReactDOM.render(<App history={browserHistory} />, document.getElementById('app'));
